pyproject.toml
[project]
name = "conot"
version = "0.1.0"
description = "Sound acquisition pipeline for cognitive note-taking"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sounddevice>=0.5.0",
    "numpy>=2.0.0",
    "scipy>=1.14.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "rich>=13.9.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "types-pyyaml>=6.0.0",
]

# STT providers (install one or more based on hardware)
stt-gpu = [
    "faster-whisper>=1.0.0",
    "torch>=2.0.0",
]
stt-cpu = [
    "pywhispercpp>=1.0.0",
]
stt-diarization = [
    "pyannote-audio>=3.0.0",
]
stt-vad = [
    "torch>=2.0.0",  # For Silero VAD
]
# Convenience: install all STT dependencies
stt = [
    "conot[stt-gpu,stt-diarization,stt-vad]",
]
stt-full = [
    "conot[stt-gpu,stt-cpu,stt-diarization,stt-vad]",
]

[project.scripts]
conot = "conot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/conot"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=conot --cov-report=term-missing"
